requirements.txt
python-telegram-bot==20.0
vk-api==11.9.0
requests==2.26.0
sqlite3
python-dotenv==0.19.2